REFACTOR: Add source and precompiled options for Madagascar
